requirements.txt
streamlit  
langchain  
langchain-openai  
faiss-cpu  
PyPDF2  
sentence-transformers  
python-dotenv  
langchain-huggingface